@@ -1,5 +1,5 @@
 class ProfilesController < ApplicationController
-  before_filter :admin_required
+  before_action :admin_required
 
   def index
     @profiles = Profile.order(:name)
