@@ -3,7 +3,7 @@ require 'rails_helper'
 RSpec.describe QuotesController, type: :controller do
   let(:current_user) { double(User).as_null_object }
   let(:current_profile) { double(Profile).as_null_object }
-  let(:quote) { double(Quote).as_null_object }
+  let(:quote) { instance_double(Quote, id: 1).as_null_object }
 
   before do
     session[:user] = 1
@@ -16,11 +16,6 @@ RSpec.describe QuotesController, type: :controller do
     it "finds all quotes belonging to the current profile" do
       expect(current_profile).to receive(:quotes)
       get "index"
-    end
-
-    it "assigns @quotes" do
-      get "index"
-      expect(assigns[:quotes]).to be
     end
   end
 
@@ -36,12 +31,12 @@ RSpec.describe QuotesController, type: :controller do
 
     it "creates a new quote" do
       expect(Quote).to receive(:new)
-      post "create", {:title => ''}
+      post "create", params: {title: ''}
     end
 
     it "assigns the current profile to the quote" do
       expect(quote).to receive(:profile_id=).with(current_profile.id)
-      post "create", {:title => ''}
+      post "create", params: {title: ''}
     end
   end
 
@@ -52,7 +47,7 @@ RSpec.describe QuotesController, type: :controller do
 
     it "finds the quote" do
       expect(Quote).to receive(:find).with("1")
-      get "preview", :id => "1"
+      get "preview", params: {id: "1"}
     end
   end
 
@@ -67,13 +62,13 @@ RSpec.describe QuotesController, type: :controller do
 
     it "finds the quote" do
       expect(Quote).to receive(:find).with("1")
-      post(:send_quote, :id => "1")
+      post(:send_quote, params:{id: "1"})
     end
 
     it "emails the quote" do
       expect(Emailer).to receive(:quote).with(quote, anything(), anything())
       expect(emailer).to receive(:deliver)
-      post(:send_quote, :id => "1")
+      post(:send_quote, params:{id: "1"})
     end
   end
 
@@ -84,16 +79,16 @@ RSpec.describe QuotesController, type: :controller do
 
     it "finds the quote" do
       expect(Quote).to receive(:find).with("1")
-      delete "destroy", :id => "1"
+      delete "destroy", params: {id: "1"}
     end
 
     it "redirects to the quotes index" do
-      delete "destroy", :id => "1"
+      delete "destroy", params: {id: "1"}
       expect(response).to redirect_to quotes_path
     end
 
     it "sets a flash[:notice] message" do
-      delete "destroy", :id => "1"
+      delete "destroy", params: {id: "1"}
       expect(flash[:notice]).to eq "Quote deleted."
     end
   end
