@@ -14,7 +14,7 @@ RSpec.describe ItemsController, type: :controller do
       it 'updates the given item' do
         item = Item.create!(name: 'Widget', information: '')
         new_name = SecureRandom.hex
-        patch :update, id: item.id, item: { name: new_name }
+        patch :update, params: {id: item.id, item: {name: new_name}}
         expect(Item.find(item.id).name).to eq new_name
       end
     end
